pyproject.toml
[project]
name = "pixel-audio-visualizer"
version = "0.0.1"
description = "Realtime audio visualizer"
authors = [
    {name = "Tolga Tarhan", email = "ttarhan@users.noreply.github.com"},
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "pyaudio>=0.2.14",
    "matplotlib>=3.9.2",
    "numpy>=2.1.1",
    "scipy>=1.14.1",
    "sacn>=1.10.0",
    "pyzmq>=26.2.0",
]

[project.scripts]
visualizercli = "visualizer.__main__:main"

[build-system]
requires = ["setuptools >= 75.3"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
packages = ["visualizer"]
